EVEREST-1208-upgrade: refactoring upgrade modal
