+++ ui/apps/everest/src/components/db-version/db-version.tsx
@@ -1,0 +1,37 @@
+// everest
+// Copyright (C) 2023 Percona LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { DBVersionFields } from './db-version.types';
+import { MenuItem } from '@mui/material';
+import { SelectInput, SelectInputProps } from '@percona/ui-lib';
+import { Messages } from './db-version.messages';
+import { DbEngineTool } from '../../shared-types/dbEngines.types';
+
+interface DbVersionProps {
+  selectInputProps?: Omit<SelectInputProps, 'name'>;
+  availableVersions: DbEngineTool[] | undefined;
+}
+export const DbVersion = ({ availableVersions }: DbVersionProps) => {
+  return (
+    <SelectInput name={DBVersionFields.dbVersion} label={Messages.dbVersion}>
+      {availableVersions &&
+        availableVersions.map((version) => (
+          <MenuItem value={version.version} key={version.version}>
+            {version.version}
+          </MenuItem>
+        ))}
+    </SelectInput>
+  );
+};
