+++ ui/apps/everest/src/components/db-version/index.ts
@@ -1,0 +1,1 @@
+export { DbVersion } from './db-version';
