+++ ui/apps/everest/src/components/db-version/utils.test.ts
@@ -1,3 +1,18 @@
+// everest
+// Copyright (C) 2023 Percona LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import { DbEngine, DbEngineType } from 'shared-types/dbEngines.types';
 import { filterAvailableDbVersionsForDbEngineEdition } from './utils';
 
@@ -12,7 +27,7 @@ const generateDbEngineWithVersions = (
     type,
   }) as unknown as DbEngine;
 
-describe('Wizard first step::utils', () => {
+describe('DBVersion Available filter test', () => {
   describe('filterAvailableDbVersionsForDbEngineEdition', () => {
     it('should rule out downgrades', () => {
       expect(
