--- ui/apps/everest/src/pages/db-cluster-details/cluster-overview/cards/db-details/basic-information/upgrade-modal.messages.ts
@@ -1,7 +1,0 @@
-export const Messages = {
-  title: 'Upgrade DB version',
-  description:
-    'Upgrading the database version updates the software to the latest release, providing new features, improved performance, enhanced security, and bug fixes. This ensures better efficiency and security for your database operations.',
-  upgrade: 'Upgrade',
-  dbVersion: 'DB version',
-};
