+++ ui/apps/everest/src/pages/db-cluster-details/dbCluster.context.types.ts
@@ -1,6 +1,10 @@
 import { DbCluster } from 'shared-types/dbCluster.types';
+import { QueryObserverResult, RefetchOptions } from '@tanstack/react-query';
 
 export interface DbClusterContextProps {
   dbCluster?: DbCluster;
   isLoading: boolean;
+  refetch: (
+    options?: RefetchOptions
+  ) => Promise<QueryObserverResult<DbCluster, unknown>>;
 }
